atax

+++ examples/polybench/atax.py
@@ -6,7 +6,7 @@ import json
 import pytest
 import allo
 import numpy as np
-from allo.ir.types import int32, float32
+from allo.ir.types import int32
 import allo.ir.types as T
 
 
@@ -18,21 +18,21 @@ def atax_np(A, x):
 
 def atax(concrete_type, m, n):
     def stage_M[
-        T: (float32, int32), M: int32, N: int32
+        T: (int32, int32), M: int32, N: int32
     ](A: "T[M, N]", x: "T[N]", out_Ax: "T[M]"):
         for m in allo.grid(M):
             for r in allo.reduction(N):
                 out_Ax[m] += A[m, r] * x[r]
 
     def stage_N[
-        T: (float32, int32), M: int32, N: int32
+        T: (int32, int32), M: int32, N: int32
     ](A: "T[M, N]", out_Ax: "T[M]", y: "T[N]"):
         for n in allo.grid(N):
             for k in allo.reduction(M):
                 y[n] += A[k, n] * out_Ax[k]
 
     def kernel_atax[
-        T: (float32, int32), M: int32, N: int32
+        T: (int32, int32), M: int32, N: int32
     ](A: "T[M, N]", x: "T[N]", y: "T[N]"):
         out_Ax: T[M] = 0
         stage_M[T, M, N](A, x, out_Ax)
@@ -64,12 +64,12 @@ def test_atax():
     test_psize = "small"
     M = psize["atax"][test_psize]["M"]
     N = psize["atax"][test_psize]["N"]
-    concrete_type = float32
+    concrete_type = int32
     sch = atax(concrete_type, M, N)
     mod = sch.build()
-    A = np.random.rand(M, N).astype(np.float32)
-    x = np.random.rand(N).astype(np.float32)
-    y = np.zeros((N,), dtype=np.float32)
+    A = np.random.rand(M, N).astype(np.int32)
+    x = np.random.rand(N).astype(np.int32)
+    y = np.zeros((N,), dtype=np.int32)
     y_ref = atax_np(A, x)
     mod(A, x, y)
     assert np.allclose(y, y_ref, atol=1e-5, rtol=1e-3)
